using avc dataset training

+++ label_sample_data.py
@@ -4,7 +4,7 @@ from PIL import Image
 from pathlib import Path
 
 def create_label(image_path, label_path):
-    arr = np.asarray(Image.open(image_path))
+    arr = np.asarray(Image.open(image_path),dtype=np.int16)
 
     # There may be a better way to do it, but this is what I have found so far
     cords = list(features.shapes(arr, mask=(arr >0)))[0][0]['coordinates'][0]
@@ -18,3 +18,5 @@ for images_dir_path in [Path(f'datasets/{x}/images') for x in ['train', 'val', '
     for img_path in images_dir_path.iterdir():
         label_path = img_path.parent.parent / 'labels' / f'{img_path.stem}.txt'
         label_line = create_label(img_path, label_path)
+
+
